+++ R/IntCal.R
@@ -6,17 +6,45 @@
 #' @docType package
 #' @author Maarten Blaauw <maarten.blaauw@qub.ac.uk> 
 #' @importFrom utils read.csv read.table write.table packageName
-#' @importFrom stats approx
+#' @importFrom stats approx dnorm
 #' @importFrom grDevices rgb
-#' @importFrom graphics legend lines points polygon segments
+#' @importFrom graphics axis par legend lines points polygon segments
 #' @name IntCal
 NULL
 
-# todo: migrate caldist and hpd from clam?
+# todo: add vignettes
+
+
+
+#' @name list.ccurves
+#' @title List the calibration curves.
+#' @description List the file names of the calibration curves available within the IntCal package.
+#' @export
+list.ccurves <- function() {
+  cc <- system.file("extdata/", package='IntCal')
+  message(cc)
+  list.files(cc)
+}
+
 
 
 #' @name copyCalibrationCurve
-#' @title Copy a calibration curve.
+#' @title Copy a calibration curve
+#' @description Copy one of the calibration curves into memory. Renamed to ccurve, and copyCalibrationCurve will become obsolete
+#' @details Copy the radiocarbon calibration curve defined by cc into memory.
+#' @return The calibration curve (invisible).
+#' @param cc Calibration curve for 14C dates: \code{cc=1} for IntCal20 (northern hemisphere terrestrial), \code{cc=2} for Marine20 (marine),
+#' \code{cc=3} for SHCal20 (southern hemisphere terrestrial). Alternatively, one can also write, e.g., "IntCal20", "Marine13".
+#' @param postbomb Use \code{postbomb=TRUE} to get a postbomb calibration curve (default \code{postbomb=FALSE}).
+#' @export
+copyCalibrationCurve <- function(cc=1, postbomb=FALSE) {
+  message("This function will be removed in future versions, as it has been renamed to ccurve.")
+  ccurve(cc, postbomb)
+}
+
+
+#' @name ccurve
+#' @title Copy a calibration curve
 #' @description Copy one of the calibration curves into memory.
 #' @details Copy the radiocarbon calibration curve defined by cc into memory.
 #' @return The calibration curve (invisible).
@@ -24,9 +52,11 @@ NULL
 #' \code{cc=3} for SHCal20 (southern hemisphere terrestrial). Alternatively, one can also write, e.g., "IntCal20", "Marine13".
 #' @param postbomb Use \code{postbomb=TRUE} to get a postbomb calibration curve (default \code{postbomb=FALSE}).
 #' @examples
-#' intcal20 <- copyCalibrationCurve(1)
-#' marine98 <- copyCalibrationCurve("Marine98")
-#' pb.sh3 <- copyCalibrationCurve("sh3", postbomb=TRUE)
+#' intcal20 <- ccurve(1)
+#' marine20 <- ccurve(2)
+#' shcal20 <- ccurve(3)
+#' marine98 <- ccurve("Marine98")
+#' pb.sh3 <- ccurve("sh3")
 #' @references
 #' Hammer and Levin 2017, "Monthly mean atmospheric D14CO2 at Jungfraujoch and Schauinsland from 1986 to 2016", heiDATA: Heidelberg Research Data Repository V2 \doi{10.11588/data/10100} 
 #'
@@ -50,7 +80,7 @@ NULL
 #'
 #' Stuiver et al. 1998 INTCAL98 radiocarbon age calibration, 24,000–0 cal BP. Radiocarbon 40, 1041-1083. \doi{10.1017/S0033822200019123}
 #' @export
-copyCalibrationCurve <- function(cc=1, postbomb=FALSE) {
+ccurve <- function(cc=1, postbomb=FALSE) {
   if(postbomb) {
     if(cc==1 || tolower(cc) == "nh1")
       fl <- "postbomb_NH1.14C" else
@@ -94,7 +124,23 @@ copyCalibrationCurve <- function(cc=1, postbomb=FALSE) {
                         fl <- "3Col_intcal98.14C" else
                         if(tolower(cc) == "marine98")
                           fl <- "3Col_marine98.14C" else  
-                            stop("cannot find this curve", call.=FALSE)
+                          if(tolower(cc) == "nh1")
+                            fl <- "postbomb_NH1.14C" else
+                            if(tolower(cc) == "nh2")
+                              fl <- "postbomb_NH2.14C" else
+                              if(tolower(cc) == "nh3")
+                                fl <- "postbomb_NH3.14C" else
+                                if(tolower(cc) == "sh1-2")
+                                  fl <- "postbomb_SH1-2.14C" else
+                                  if(tolower(cc) == "sh3")
+                                    fl <- "postbomb_SH3.14C" else
+                                    if(tolower(cc) == "kure")
+                                      fl <- "kure.14C" else
+                                      if(tolower(cc) == "levinkromer")
+                                        fl <- "LevinKromer.14C" else
+                                        if(tolower(cc) == "santos")
+                                          fl <- "Santos.14C" else
+                                          stop("cannot find this curve", call.=FALSE)
   cc <- system.file("extdata/", fl, package='IntCal')
   cc <- read.table(cc)
   invisible(cc)
@@ -102,14 +148,16 @@ copyCalibrationCurve <- function(cc=1, postbomb=FALSE) {
 
 
 
-#' @name draw.calibrationcurve
-#' @title Draw an IntCal calibration curve.
+#' @name draw.ccurve
+#' @title Draw a calibration curve.
 #' @description Draw one or two of the calibration curves, or add a calibration curve to an existing plot.
 #' @return The calibration curve (invisible).
 #' @param cal1 First calendar year for the plot
 #' @param cal2 Last calendar year for the plot
-#' @param cc1 Name of the calibration curve. Can be "IntCal20", "Marine20", "SHCal20", or for the previous curves "IntCal13", "Marine13" or "SHCal13".
+#' @param cc1 Name of the calibration curve. Can be "IntCal20", "Marine20", "SHCal20", or for the previous curves "IntCal13", "Marine13" or "SHCal13". Can also be "nh1", "nh2", "nh3", "sh1-2", "sh3", "Kure", "LevinKromer" or "Santos" for postbomb curves.
 #' @param cc2 Optional second calibration curve to plot. Can be "IntCal20", "Marine20", "SHCal20", or for the previous curves "IntCal13", "Marine13" or "SHCal13". Defaults to nothing, NA.
+#' @param cc1.postbomb Use \code{postbomb=TRUE} to get a postbomb calibration curve for cc1 (default \code{cc1.postbomb=FALSE}).
+#' @param cc2.postbomb Use \code{postbomb=TRUE} to get a postbomb calibration curve for cc2 (default \code{cc2.postbomb=FALSE}).
 #' @param BCAD The calendar scale of graphs and age output-files is in cal BP (calendar or calibrated years before the present, where the present is AD 1950) by default, but can be changed to BC/AD using \code{BCAD=TRUE}.
 #' @param cal.lab The labels for the calendar axis (default \code{age.lab="cal BP"} or \code{"BC/AD"} if \code{BCAD=TRUE}), or to \code{age.lab="kcal BP"} etc. if ka=TRUE.
 #' @param cal.rev Reverse the calendar axis. 
@@ -117,6 +165,7 @@ copyCalibrationCurve <- function(cc=1, postbomb=FALSE) {
 #' @param c14.lim Axis limits for the C-14 axis. Calculated automatically by default. 
 #' @param c14.rev Reverse the C-14 axis.
 #' @param ka Use kcal BP (and C14 kBP).
+#' @param add.yaxis Whether or not to plot the second calibration. Defaults to \code{add.yaxis=FALSE}.
 #' @param cc1.col Colour of the calibration curve (outline).
 #' @param cc1.fill Colour of the calibration curve (fill).
 #' @param cc2.col Colour of the calibration curve (outline), if activated (default cc2=NA).
@@ -125,13 +174,15 @@ copyCalibrationCurve <- function(cc=1, postbomb=FALSE) {
 #' @param bty Draw a box around a box of a certain shape. Defaults to \code{bty="l"}.
 #' @param ... Any additional optional plotting parameters. 
 #' @examples 
-#' draw.calibrationcurve()
-#' draw.calibrationcurve(1000, 3000, cc2="Marine20")
+#' draw.ccurve()
+#' draw.ccurve(1000, 3000, cc2="Marine20")
+#' draw.ccurve(1800, 2020, BCAD=TRUE, cc2="nh1", cc2.postbomb=TRUE)
+#' draw.ccurve(1800, 2010, BCAD=TRUE, cc2="nh1", add.yaxis=TRUE)
 #' @export
-draw.calibrationcurve <- function(cal1=-50, cal2=55e3, cc1="IntCal20", cc2=NA, BCAD=FALSE, cal.lab=NA, cal.rev=FALSE, c14.lab=NA, c14.lim=NA, c14.rev=FALSE, ka=FALSE, cc1.col=rgb(0,0,1,.5), cc1.fill=rgb(0,0,1,.2), cc2.col=rgb(0,.5,0,.5), cc2.fill=rgb(0,.5,0,.2), add=FALSE, bty="l", ...) {
+draw.ccurve <- function(cal1=-50, cal2=55e3, cc1="IntCal20", cc2=NA, cc1.postbomb=FALSE, cc2.postbomb=FALSE,  BCAD=FALSE, cal.lab=NA, cal.rev=FALSE, c14.lab=NA, c14.lim=NA, c14.rev=FALSE, ka=FALSE, add.yaxis=FALSE, cc1.col=rgb(0,0,1,.5), cc1.fill=rgb(0,0,1,.2), cc2.col=rgb(0,.5,0,.5), cc2.fill=rgb(0,.5,0,.2), add=FALSE, bty="l", ...) {
 
   # read and narrow down the calibration curve(s)
-  cc.1 <- copyCalibrationCurve(cc1)
+  cc.1 <- ccurve(cc1, cc1.postbomb)
   if(BCAD)
     cc.1[,1] <- 1950 - cc.1[,1] 
   mindat <- cc.1[,1] >= min(cal1, cal2)
@@ -142,7 +193,7 @@ draw.calibrationcurve <- function(cal1=-50, cal2=55e3, cc1="IntCal20", cc2=NA, B
   cc1.pol <- cbind(c(cc.1[,1], rev(cc.1[,1])), c(cc.1[,2]-cc.1[,3], rev(cc.1[,2]+cc.1[,3])))
   
   if(!is.na(cc2)) {
-    cc.2 <- copyCalibrationCurve(cc2)
+    cc.2 <- ccurve(cc2, cc2.postbomb)
     if(BCAD)
       cc.2[,1] <- 1950 - cc.2[,1] 
     mindat <- cc.2[,1] >= min(cal1, cal2)
@@ -176,7 +227,7 @@ draw.calibrationcurve <- function(cal1=-50, cal2=55e3, cc1="IntCal20", cc2=NA, B
     }
 
     if(is.na(c14.lim))
-      if(is.na(cc2))
+      if(is.na(cc2) || add.yaxis)
         c14.lim <- range(cc1.pol[,2]) else
           c14.lim <- range(cc1.pol[,2], cc2.pol[,2])
     if(c14.rev)
@@ -186,14 +237,22 @@ draw.calibrationcurve <- function(cal1=-50, cal2=55e3, cc1="IntCal20", cc2=NA, B
     plot(0, type="n", xlim=cal.lim, xlab=callab, ylim=c14.lim, ylab=c14lab, bty=bty, ...)
   }
 
-  # add the calibration curve(s)
+  # add the calibration curve
   polygon(cc1.pol, col=cc1.fill, border=NA) # calibration curve
   lines(cc.1[,1], cc.1[,2]-cc.1[,3], col=cc1.col)
   lines(cc.1[,1], cc.1[,2]+cc.1[,3], col=cc1.col)
+
+  # add a second curve?
   if(!is.na(cc2)) {
+    if(add.yaxis) {
+      op <- par(new=TRUE)
+      plot(cc2.pol, type="n", xlim=cal.lim, xlab="", ylab="", bty="n", xaxt="n", yaxt="n")
+    }
     polygon(cc2.pol, col=cc2.fill, border=NA) # calibration curve
     lines(cc.2[,1], cc.2[,2]-cc.2[,3], col=cc2.col)
-    lines(cc.2[,1], cc.2[,2]+cc.2[,3], col=cc2.col)	
+    lines(cc.2[,1], cc.2[,2]+cc.2[,3], col=cc2.col)
+    if(add.yaxis)
+      axis(4, col=cc2.col, col.axis=cc2.col)
   }
 }
 
@@ -483,7 +542,7 @@ age.pMC <- function(mn, sdev, ratio=100, decimals=3) {
 #' @export
 intcal.data <- function(cal1, cal2, cc1="IntCal20", cc2=NA, calcurve.data="IntCal20", BCAD=FALSE, cal.lab=NA, cal.rev=FALSE, c14.lab=NA, c14.lim=NA, c14.rev=FALSE, ka=FALSE, cc1.col=rgb(0,0,1,.5), cc1.fill=rgb(0,0,1,.2), cc2.col=rgb(0,.5,0,.5), cc2.fill=rgb(0,.5,0,.2), data.cols=1:8, data.pch=c(1,2,5,6,15:19), pch.cex=.5, legend.loc="topleft", legend.ncol=2, legend.cex=0.7, cc.legend="bottomright", bty="l",  ...) {
 
-  # read the data. 
+  # read the data
   if(tolower(calcurve.data) == "intcal20") {
     dat <- system.file("extdata/", "intcal20_data.txt", package = "IntCal")
     sourcesdat <- system.file("extdata/", "intcal20_data_sources.txt", package = "IntCal")
@@ -495,8 +554,6 @@ intcal.data <- function(cal1, cal2, cc1="IntCal20", cc2=NA, calcurve.data="IntCa
     sourcesdat <- read.table(sourcesdat, nrows=9, sep=",")
   }
   dat <- read.table(dat, header=TRUE, sep=" ")
-  dat <<- dat
-  sourcesdat <<- sourcesdat
 
   if(BCAD)
     dat$cal <- 1950 - dat$cal
@@ -507,7 +564,7 @@ intcal.data <- function(cal1, cal2, cc1="IntCal20", cc2=NA, calcurve.data="IntCa
   dat <- dat[which( mindat * maxdat == 1),]
 
   # read and narrow down the calibration curve(s)
-  cc.1 <- copyCalibrationCurve(cc1)
+  cc.1 <- ccurve(cc1)
   if(BCAD)
     cc.1[,1] <- 1950 - cc.1[,1] 
   mindat <- cc.1[,1] >= min(cal1, cal2)
@@ -518,7 +575,7 @@ intcal.data <- function(cal1, cal2, cc1="IntCal20", cc2=NA, calcurve.data="IntCa
   cc1.pol <- cbind(c(cc.1[,1], rev(cc.1[,1])), c(cc.1[,2]-cc.1[,3], rev(cc.1[,2]+cc.1[,3])))
   
   if(!is.na(cc2)) {
-    cc.2 <- copyCalibrationCurve(cc2)
+    cc.2 <- ccurve(cc2)
     if(BCAD)
       cc.2[,1] <- 1950 - cc.2[,1] 
     mindat <- cc.2[,1] >= min(cal1, cal2)
@@ -533,7 +590,6 @@ intcal.data <- function(cal1, cal2, cc1="IntCal20", cc2=NA, calcurve.data="IntCa
   sets <- dat$set
   set.cols <- NA; set.pchs <- NA
   these.sets <- sort(unique(sets))
-  these.sets <<- these.sets
   these.cols <- data.cols[1:length(sets)]
   these.pchs <- data.pch[1:length(sets)]
   for(i in 1:length(these.sets)) {